No longer can you specify the underlying value type for stats. I originally did this to make it so you could choose the value type you needed for each stat on a per-stat basis (no point in having an int for a level, for example), but turns out they were all being boxed. So in attempt to save 2-3 bytes, I ended up costing an extra 10 bytes plus the boxing/unboxing overhead.

+++ netgore/trunk/DemoGame.Server/World/Map/Entities/Characters/Character.cs
@@ -516,13 +516,13 @@ namespace DemoGame.Server
             ModStats.GetStat(StatType.MaxHP).Changed += delegate(IStat<StatType> stat)
             {
                 if (HP > stat.Value)
-                    HP = stat.Value;
+                    HP = (int)stat.Value;
             };
 
             ModStats.GetStat(StatType.MaxMP).Changed += delegate(IStat<StatType> stat)
             {
                 if (MP > stat.Value)
-                    MP = stat.Value;
+                    MP = (int)stat.Value;
             };
         }
